requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
fuzzywuzzy==0.18.0
gunicorn==21.2.0
idna==3.6
jellyfish==1.0.3
Levenshtein==0.23.0
mysqlclient==2.2.1
natsort==8.4.0
numpy==1.26.3
opencv-python==4.9.0.80
packaging==23.2
pandas==2.1.4
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.3
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
pypdfium2==4.26.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-Levenshtein==0.23.0
pytz==2023.3.post1
rapidfuzz==3.6.1
requests==2.31.0
scipy==1.11.4
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
whitenoise==6.6.0
